backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
opencv-python==4.9.0.80
scikit-learn==1.3.2 